ctrl_dbus: fix cross compile build (commend)
On newer systems `gdbus-codegen` generates `baresipbus.c` with

```c
  #if GLIB_VERSION_MAX_ALLOWED >= GLIB_VERSION_2_84
    g_variant_builder_init_static (&builder, G_VARIANT_TYPE ("a{sv}"));
  #else
    g_variant_builder_init(&builder, G_VARIANT_TYPE ("a{sv}"));
  #endif
```

But GLIB_VERSION_2_84 is not defined if a cross-compiler toolchain with
an older glib version is used. Then the #if condition is true which
leads to a build error because `g_variant_builder_init_static` is not
defined.

(commend) status: Because it should not be a general problem. Maybe we
use the wrong gdbus-codegen.